fix: updated heights and compression data types

+++ backend/app/services/experiments/models.py
@@ -20,14 +20,14 @@ class ExperimentBase(SQLModel):
     applied_excitation_directions: List[str] | None = Field(
         sa_column=Column(ARRAY(String)), default=None)
     storeys_nb: int | None = Field(default=None)
-    building_height: int | None = Field(default=None)
-    total_building_height: int | None = Field(default=None)
+    building_height: float | None = Field(default=None)
+    total_building_height: float | None = Field(default=None)
     diaphragm_material: str | None = Field(default=None)
     roof_material_geometry: str | None = Field(default=None)
     masonry_unit_type: str | None = Field(default=None)
     masonry_unit_material: str | None = Field(default=None)
     mortar_type: str | None = Field(default=None)
-    masonry_compressive_strength: int | None = Field(default=None)
+    masonry_compressive_strength: float | None = Field(default=None)
     masonry_wall_thickness: List[int] | None = Field(
         sa_column=Column(ARRAY(Integer)), default=None)
     wall_leaves_nb: int | None = Field(default=None)
